Add UserDefaultsLocker
